moved Exercises data to lib
